+++ sebimachine/shared_libs/aliases.json
@@ -1,0 +1,6 @@
+{
+    "sar": "sar",
+    "selfrole": "sar",
+    "selfroles": "sar"
+
+}
