+++ sebimachine/shared_libs/loggable.py
