--- src/shared_libs/aliases.json
@@ -1,6 +1,0 @@
-{
-    "sar": "sar",
-    "selfrole": "sar",
-    "selfroles": "sar"
-
-}
